+++ CHANGELOG.md
@@ -33,9 +33,15 @@ Security updates
 
 ## [1.1.1] - 2026-05-29
 
-### Added
+### Changed
 
-- Test123
+- UI: Move search results out of search box and show total value of currently shown items
+
+### Fixed
+
+- Fix when starting app without network, it hangs on loading screen.
+- Fix when changing accounts while a search or filter is active, it does not reset the filter.
+- Fix when transferring items, need to select container as well or else "Transfer" button will stay disabled.
 
 ## [1.1.0] - 2026-01-28
 
@@ -81,7 +87,7 @@ Security updates
 - Initial release of SkinsTrackr
 
 [unreleased]: https://github.com/SkinsTrackr/skinstrackr/compare/v1.1.1...HEAD
-[1.1.0]: https://github.com/SkinsTrackr/skinstrackr/releases/tag/v1.1.1
+[1.1.1]: https://github.com/SkinsTrackr/skinstrackr/releases/tag/v1.1.1
 [1.1.0]: https://github.com/SkinsTrackr/skinstrackr/releases/tag/v1.1.0
 [1.0.2]: https://github.com/SkinsTrackr/skinstrackr/releases/tag/v1.0.2
 [1.0.1]: https://github.com/SkinsTrackr/skinstrackr/releases/tag/v1.0.1